Remove MusigClient. Rename and move mock server
